+++ lib/Screens/Login/components/already_have_an_account.dart
@@ -1,0 +1,33 @@
+import 'package:flutter/material.dart';
+import 'package:picki/constant.dart';
+
+
+class AlreadyHaveAccountWidget extends StatelessWidget {
+  final bool login;
+  final Function press;
+  const AlreadyHaveAccountWidget({
+    Key key,
+    this.login = true,
+    this.press,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Row(mainAxisAlignment: MainAxisAlignment.center, children: <Widget>[
+      Text(
+        login ? "Dont have an Account?":"Already have an Account?",
+        style: TextStyle(color: kPrimaryColor),
+      ),
+      GestureDetector(
+        onTap: press,
+        child: Text(
+         login ?  "Sign Up": "Sign In",
+          style: TextStyle(
+            color: kPrimaryColor,
+            fontWeight: FontWeight.bold,
+          ),
+        ),
+      ),
+    ]);
+  }
+}
